+++ plugins/example-plugin/src/routes/api/example/mod.rs
@@ -1,2 +1,0 @@
-//! Example routes
-pub mod items;